Move ThreadQueueList to a separate file.
Not reused anywhere, but just so it's a little cleaner.
